+++ indexer-api/src/infra/api/v1/subscription/unshielded.rs
@@ -16,6 +16,7 @@ use crate::{
     infra::api::{
         ApiError, ApiResult, ContextExt, ResultExt,
         v1::{
+            subscription::get_next_transaction,
             transaction::Transaction,
             unshielded::{UnshieldedAddress, UnshieldedUtxo},
         },
@@ -23,7 +24,7 @@ use crate::{
 };
 use async_graphql::{Context, SimpleObject, Subscription, Union, async_stream::try_stream};
 use derive_more::Debug;
-use fastrace::{Span, future::FutureExt, prelude::SpanContext, trace};
+use fastrace::trace;
 use futures::{Stream, StreamExt, TryStreamExt};
 use indexer_common::domain::{NetworkId, RawUnshieldedAddress, Subscriber, UnshieldedUtxoIndexed};
 use log::{debug, warn};
@@ -283,16 +284,4 @@ where
     Ok(UnshieldedTransactionsProgress {
         highest_transaction_id,
     })
-}
-
-async fn get_next_transaction<E>(
-    transactions: &mut (impl Stream<Item = Result<domain::Transaction, E>> + Unpin),
-) -> Result<Option<domain::Transaction>, E> {
-    transactions
-        .try_next()
-        .in_span(Span::root(
-            "transactions-subscription.get-next-transaction",
-            SpanContext::random(),
-        ))
-        .await
 }
